Add some optimisation of Netlify
Add a gulp task for Netlify minified builds.

Add some HTTP/2 headers for pushing CSS and JS to clients in the initial request.
This isn't compatible with the netlify bundling as that uses a custom CloudFront distribution with a hash name. We cannot know this to put in the headers file.
Trade off with bundling for pushing benefit, may not be worth it.